aider: Added WaitingForPlayers state to AppState enum, modified button_system to set AppState to WaitingForPlayers when OnlinePlayButton is clicked, and added show_waiting_text system to display "Waiting for players" text when AppState is WaitingForPlayers.
# Context:
USER: Lets create a WaitingForPlayers AppState this way once a person clicks play online it then transition to this new `AppState`. Once this state has enter lets show a Waiting for players text in the middle of the window.
USER: Lets create a WaitingForPlayers AppState this way once a person clicks play online it then transition to this new `AppState`. Once this state has enter lets show a Waiting for players text in the middle of the window.
ASSISTANT: To implement this, we need to modify the `src\app_state.rs`, `src\menu.rs` and `src\main.rs` files.

First, we will add the `WaitingForPlayers` state to the `AppState` enum in `src\app_state.rs`:

```rust
src\app_state.rs
<<<<<<< ORIGINAL
#[derive(Clone, Eq, PartialEq, Debug, Hash)]
pub enum AppState {
    Menu,
    Online,
}
=======
#[derive(Clone, Eq, PartialEq, Debug, Hash)]
pub enum AppState {
    Menu,
    Online,
    WaitingForPlayers,
}
>>>>>>> UPDATED
```

Then, we will modify the `button_system` function in `src\menu.rs` to change the `AppState` to `WaitingForPlayers` when the `OnlinePlayButton` is clicked:

```rust
src\menu.rs
<<<<<<< ORIGINAL
fn button_system(
    mut interaction_query: Query<
        (&Interaction, &MenuButtons),
        (Changed<Interaction>, With<Button>),
    >,
    mut app_state: ResMut<NextState<AppState>>,
) {
    // Handle button interactions
    for (interaction, button) in interaction_query.iter_mut() {
        match *interaction {
            Interaction::Clicked => {
                match button {
                    MenuButtons::OnlinePlay => {
                        app_state.set(AppState::Online);
                    }
                    _ => {}
                }
            }
            _ => {}
        }
    }
}
=======
fn button_system(
    mut interaction_query: Query<
        (&Interaction, &MenuButtons),
        (Changed<Interaction>, With<Button>),
    >,
    mut app_state: ResMut<NextState<AppState>>,
) {
    // Handle button interactions
    for (interaction, button) in interaction_query.iter_mut() {
        match *interaction {
            Interaction::Clicked => {
                match button {
                    MenuButtons::OnlinePlay => {
                        app_state.set(AppState::WaitingForPlayers);
                    }
                    _ => {}
                }
            }
            _ => {}
        }
    }
}
>>>>>>> UPDATED
```

Finally, we will modify the `main` function in `src\main.rs` to add the `WaitingForPlayers` state and a system that displays a "Waiting for players" text when the `AppState` is `WaitingForPlayers`:

```rust
src\main.rs
<<<<<<< ORIGINAL
fn main() {
    let mut app = App::new();
    app
        .add_state::<AppState>()
        .add_plugins(DefaultPlugins)
        .add_plugin(CameraPlugin)
        .add_plugin(NetworkPlugin)
        .add_plugin(MenuPlugin)
        .run();
}
=======
fn main() {
    let mut app = App::new();
    app
        .add_state(AppState::Menu)
        .add_plugins(DefaultPlugins)
        .add_plugin(CameraPlugin)
        .add_plugin(NetworkPlugin)
        .add_plugin(MenuPlugin)
        .add_system_set(SystemSet::on_enter(AppState::WaitingForPlayers).with_system(show_waiting_text.system()))
        .run();
}

fn show_waiting_text(mut commands: Commands, asset_server: Res<AssetServer>) {
    let font = asset_server.load("Roboto/Roboto-Regular.ttf");
    commands.spawn(TextBundle {
        text: Text {
            value: "Waiting for players...".to_string(),
            font: font,
            style: TextStyle {
                font_size: 50.0,
                color: Color::WHITE,
                ..Default::default()
            },
        },
        ..Default::default()
    });
}
>>>>>>> UPDATED
```

This code adds a new system `show_waiting_text` that spawns a text entity with the value "Waiting for players..." when the `AppState` is `WaitingForPlayers`. The `button_system` function in `src\menu.rs` is modified to set the `AppState` to `WaitingForPlayers` when the `OnlinePlayButton` is clicked. The `AppState::WaitingForPlayers` state is added to the `AppState` enum in `src\app_state.rs`.