aider: Changed the `MyGameInput` struct to use the `#[repr(C, packed)]` attribute to ensure the same layout as in C and that it is packed, fixing the error.
# Context:
USER: I'm getting an error on my Code:error[E0512]: cannot transmute between types of different sizes, or dependently-sized types
  --> src\input.rs:19:12
   |
19 | pub struct MyGameInput {
   |            ^^^^^^^^^^^
   |
   = note: source type: `MyGameInput` (96 bits)
   = note: target type: `TypeWithoutPadding` (72 bits)
ASSISTANT: The error message is indicating that the `MyGameInput` struct is being transmuted to a type of a different size. This is likely due to the `Pod` and `Zeroable` traits from the `bytemuck` crate, which require the types to be Plain Old Data (POD) and able to be safely set to zero, respectively.

The `MyGameInput` struct has two `f32` fields and a `u8` field. The total size of these fields is 72 bits, but the struct itself is 96 bits. This discrepancy is likely due to padding added by Rust for alignment purposes.

To fix this issue, you can use the `#[repr(C, packed)]` attribute on the `MyGameInput` struct to ensure that it has the same layout as in C and that it is packed, meaning that there is no padding between fields.

Here is the change you need to make:

```rust
src\input.rs
<<<<<<< ORIGINAL
#[repr(C)]
#[derive(Copy, Clone, Pod, Zeroable, Default)]
pub struct MyGameInput {
    pub axis_lx: f32,
    pub axis_ly: f32,
    pub actions: u8,
}
=======
#[repr(C, packed)]
#[derive(Copy, Clone, Pod, Zeroable, Default)]
pub struct MyGameInput {
    pub axis_lx: f32,
    pub axis_ly: f32,
    pub actions: u8,
}
>>>>>>> UPDATED
```

Please note that using `#[repr(C, packed)]` can have performance implications because it may prevent certain kinds of hardware memory access optimizations.