aider: Refactored move_player_system to use axis inputs for player movement.
# Context:
USER: Can you refactor the move_player_system function so that it adds velocity to the player but instead of using the actions use axis for the inputs.
ASSISTANT: Sure, I can help with that. We will modify the `move_player_system` function in the `src\level.rs` file to use the `axis_lx` and `axis_ly` fields from the `MyGameInput` struct for movement instead of the `actions` field.

Here are the changes:

```rust
src\level.rs
<<<<<<< ORIGINAL
fn move_player_system(
    mut query: Query<(&mut Velocity, &mut Transform, &Player), With<Rollback>>,
    inputs: Res<PlayerInputs<GgrsConfig>>,
) {
    for (mut vel, mut transform, player) in query.iter_mut() {
        let input = inputs[player.handle].0;
        let actions = input.actions;
        let mut v = vel.0;
        if actions & INPUT_MOVE_UP != 0 && actions & INPUT_MOVE_DOWN == 0 {
            v.z -= MOVEMENT_SPEED;
        }
        if actions & INPUT_MOVE_UP == 0 && actions & INPUT_MOVE_DOWN != 0 {
            v.z += MOVEMENT_SPEED;
        }
        let mag = ComplexField::sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
        if mag > MAX_SPEED {
            let factor = MAX_SPEED / mag;
            v.x *= factor;
            v.y *= factor;
            v.z *= factor;
        }
        transform.translation += v;
    }
}
=======
fn move_player_system(
    mut query: Query<(&mut Velocity, &mut Transform, &Player), With<Rollback>>,
    inputs: Res<PlayerInputs<GgrsConfig>>,
) {
    for (mut vel, mut transform, player) in query.iter_mut() {
        let input = inputs[player.handle].0;
        let mut v = vel.0;
        v.x += input.axis_lx * MOVEMENT_SPEED;
        v.z += input.axis_ly * MOVEMENT_SPEED;
        let mag = ComplexField::sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
        if mag > MAX_SPEED {
            let factor = MAX_SPEED / mag;
            v.x *= factor;
            v.y *= factor;
            v.z *= factor;
        }
        transform.translation += v;
    }
}
>>>>>>> UPDATED
```

In the updated code, we are adding the `axis_lx` and `axis_ly` values from the `MyGameInput` struct to the `x` and `z` components of the velocity, respectively. This will cause the player to move in the direction specified by the joystick axes. The speed of movement is controlled by the `MOVEMENT_SPEED` constant. The rest of the function remains the same, limiting the maximum speed and updating the player's position.